ui/requirements.txt
# AutoResearch Agent - Dependencies
# Run: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI & LLM
groq>=0.4.0
langchain>=0.1.0
langgraph>=0.0.50

# Search & Data
tavily-python>=0.3.0

# Data Processing
pydantic>=2.5.0

# Note: Ensure you have set up your .env file with:
# GROQ_API_KEY=your_key_here
# TAVILY_API_KEY=your_key_here
